Revert "added asynchronous API for SMB resource inspection in browser"
This reverts commit 22aaf31fea03f1ff9c3cfb2949413892d2da43f0.

This implementation of an asynchronous API is problematic when using multiple
worker processes with gunicorn, since the "tasks" dictionary is duplicated
for each worker, so that if one worker registers the task and the client side
asks another worker about its progress, a "404 not found" is returned. The
problem can be "visualized" by printing hex(id(tasks)) in the wrapper
function in asyncapi.py, showing that the different workers (obviously) have
different memory addresses for the "tasks" dictionary.