disable caching for AJAX requests
this fixes a bug in Internet Explorer, e.g. when reloading the sample
after adding an action, the new action is not shown / when modifying an
action, the old action is shown after saving, even though the data has
been updated on the server

c.f. https://www.itworld.com/article/2693447/ajax-requests-not-executing
-or-updating-in-internet-explorer-solution.html